# Improve creation performance of HttpIncoming