--- a/.eslintrc
+++ b/.eslintrc
@@ -14,6 +14,7 @@
     "plugins": ["prettier"],
     "root": true,
     "rules": {
+        "no-underscore-dangle": [0],
         "class-methods-use-this": [0],
         "strict": [0, "global"]
     }