`HttpIncoming` lives in hot path of every request. This PR doubles the creation performance of `HttpIncoming`.

From:

```sh
new HttpIncoming() - No params x 91,806 ops/sec ±2.48% (85 runs sampled)
new HttpIncoming() - With params x 92,816 ops/sec ±2.34% (85 runs sampled)
```

to:

```sh
new HttpIncoming() - No params x 186,691 ops/sec ±1.04% (90 runs sampled)
new HttpIncoming() - With params x 186,027 ops/sec ±1.54% (92 runs sampled)
```

There is probably more performance to press out of this but that involves taking a stab on the `original-url` module.